+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Change Log
 
-### Version 0.0.2
+### Version 0.0.3: March 11, 2018
+
+- added option for package and rule excludes
+
+### Version 0.0.2: March 10, 2018
 
 - support for OSX
 - use the right path for setting up the workspace
 - fixed wrong label decomposition
 - only consider c++ targets for c_cpp_properties.json generation
 
-### Version 0.0.1
+### Version 0.0.1: March 9, 2018
 
-- Preview release (experimental)
+- preview release (experimental)